requirements.txt
numpy
PyAudio
pyyaml
